feat: Configure multiple exits and their respective types
