--- lib/main.ts
@@ -1,12 +1,0 @@
-export * from '@/utils/Array/index'
-export * from '@/utils/Browser/index'
-export * from '@/utils/FormatDate/index'
-export * from '@/utils/FormatDate/constant'
-export * from '@/utils/HighOrderFunction/index'
-export * from '@/utils/TypeCheck/index'
-export * from '@/utils/Number/index'
-export * from '@/utils/String/index'
-export * from '@/utils/Object/index'
-export * from '@/special/WeChat/index'
-
-export type * from './types/global'
