+++ lib/public/HighOrderFunction/index.ts
