+++ lib/public/Number/index.ts
