+++ lib/public/Object/index.ts
