+++ lib/public/String/index.ts
