+++ lib/public/TypeCheck/index.ts
