+++ lib/public/types/global.d.ts
@@ -56,10 +56,4 @@ export declare namespace TYPE {
   interface anyType {
     [key: string]: any
   }
-
-  interface IWeChatUploadFile extends anyType {
-    url: string
-    name: string
-    filePath: string
-  }
 }
