+++ lib/special/Vue/main.ts
