+++ lib/special/Vue/types/global.d.ts
@@ -1,0 +1,1 @@
+export declare namespace TYPE {}
