+++ lib/special/WeChat/main.ts
@@ -1,0 +1,4 @@
+export * from './CreateQRCode/index'
+export * from './UploadFile/index'
+
+export type * from './types/global'
