+++ lib/special/WeChat/types/global.d.ts
@@ -1,0 +1,7 @@
+export declare namespace TYPE {
+  interface IWeChatUploadFile extends anyType {
+    url: string
+    name: string
+    filePath: string
+  }
+}
