+++ lib/special/WeChat/types/index.d.ts
