+++ src/App.tsx
@@ -3,7 +3,7 @@ import HelloWorld from './components/HelloWorld'
 export default function App() {
   return (
     <div>
-      <HelloWorld/>
+      <HelloWorld />
     </div>
   )
 }
