+++ src/components/Test.vue
@@ -1,5 +1,4 @@
 <script></script>
 <template>
-  <div> 后面准备编写自定义指令 所以集成了tsx 和 vue 配置多入口888
-  </div>
+  <div>后面准备编写自定义指令 所以集成了tsx 和 vue 配置多入口888</div>
 </template>
